feat: create file structure with dependencies
Includes directories for microservices, database containers, testing, documentation, and development scripting. The microservice directories include all the npm dependencies need to get starting right away and some template code for the backend.

This was creating to provide organization and a template for each contributor to start working quickly on the backend code.