fix: delete database directory structure
The initial thought was that each database needed its own dockerfile and startup scripts.
However we can make the db containers directly on docker compose and make the microservices
dependant on them. We can furthermore include the startup scripts inside the microservice,
such that when the microservice boots up, it will also boot up the database. This will make
the dbs ephemeral like the microservices, however, this will be sufficient for this project.